Switch 2 design advancement
Inner text gets changed to 1 from 0 if checked is true now. I also played with css a bit. The struggle is I want to keep it as much  ad plain text as possible, but I also want it to be obviously clickable. Have to think about it for a bit.