I keep working on the glitch effect I want to acheive.
Image source issue was due to how the live server runs and how repo is set up locally, just as expected. I was able to fix it by not running the live server. Not important for now. Glitch semiworking now. What I want to fix next is that first,  setInterval keeps running after the event is over. I also might change the svgs.